backend/requirements.txt
accelerate
blinker
bs4
fastapi
uvicorn
jinja2
gradio
huggingface-hub
langchain
langchain-ibm
langchain-community
pdf2image
chromadb
pypdf
tiktoken
pillow
ibm_watsonx_ai
ibm_watson
ibm_watson_machine_learning
tf-keras
tokenizers
torch
torchaudio
torchvision
transformers
python-dotenv
pydantic
pandas
numpy
sentence-transformers
InstructorEmbedding
p4python
lxml
